requirements.txt
telegraph==1.4.0
beautifulsoup4==4.7.1
feedparser==5.2.1
requests==2.21.0
python-telegram-bot==12.0.0b1
